% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{cal_diff_map}
\alias{cal_diff_map}
\title{Calculate difference map
This function produces a map of difference b/w mean and sum indexes in rank of cropland harvested area fraction.}
\usage{
cal_diff_map(
  mean_index_rast,
  cropharvest_aggtm_crop,
  cropharvest_agglm_crop,
  geoscale,
  resolution = the$parameters_config$`CCRI parameters`$Resolution
)
}
\arguments{
\item{mean_index_rast}{A raster object for mean index raster difference}

\item{cropharvest_aggtm_crop}{A raster object for cropland harvest}

\item{cropharvest_agglm_crop}{A raster object for cropland harvest}

\item{geoscale}{geographical scale with longitude and latitudes. It will be converted into extent.
See \code{\link[=get_geographic_scales]{get_geographic_scales()}} and \code{\link[terra:ext]{terra::ext()}}.}

\item{resolution}{resolution to plot raster and map}
}
\description{
Calculate difference map
This function produces a map of difference b/w mean and sum indexes in rank of cropland harvested area fraction.
}
