% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{calculate_ccri}
\alias{calculate_ccri}
\title{Calculate Cropland Connectivity Risk Index (CCRI)
This function calculates CCRI for given parameters using power law and negative exponential.
It returns a list of CCRI values.}
\usage{
calculate_ccri(
  link_threshold = 0,
  power_law_metrics =
    the$parameters_config$`CCRI parameters`$NetworkMetrics$InversePowerLaw,
  negative_exponential_metrics =
    the$parameters_config$`CCRI parameters`$NetworkMetrics$NegativeExponential,
  crop_cells_above_threshold,
  thresholded_crop_values
)
}
\arguments{
\item{link_threshold}{A threshold value for link}

\item{power_law_metrics}{A list of power law metrics}

\item{negative_exponential_metrics}{A list of negative exponential metrics}

\item{crop_cells_above_threshold}{A list of crop cells above threshold}

\item{thresholded_crop_values}{A list of crop values above threshold}
}
\value{
A list of calculated CCRI values
}
\description{
Calculate Cropland Connectivity Risk Index (CCRI)
This function calculates CCRI for given parameters using power law and negative exponential.
It returns a list of CCRI values.
}
