% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{calculate_difference_map}
\alias{calculate_difference_map}
\title{Calculate difference map
This function produces a map of difference in rank of cropland harvested area fraction}
\usage{
calculate_difference_map(
  mean_index_raster_diff,
  cropharvest_aggtm_crop,
  cropharvest_agglm_crop,
  zero_extent_raster,
  map_grey_background_ext,
  resolution = the$parameters_config$`CCRI parameters`$Resolution
)
}
\arguments{
\item{mean_index_raster_diff}{A raster object for mean index raster difference}

\item{cropharvest_aggtm_crop}{A raster object for cropland harvest}

\item{cropharvest_agglm_crop}{A raster object for cropland harvest}

\item{zero_extent_raster}{A raster object for zero extent raster}

\item{map_grey_background_ext}{A raster object for map grey background extent}

\item{resolution}{resolution to plot raster and map}
}
\description{
Calculate difference map
This function produces a map of difference in rank of cropland harvested area fraction
}
