% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccri_helper.R
\name{calculate_metrics_weight}
\alias{calculate_metrics_weight}
\title{calculate weights for each metric}
\usage{
calculate_metrics_weight(
  betweenness_metric = FALSE,
  node_strength = FALSE,
  sum_of_nearest_neighbors = FALSE,
  eigenvector_centrality = FALSE
)
}
\arguments{
\item{betweenness_metric}{A logical value indicating if the betweenness metric
should be used}

\item{node_strength}{A logical value indicating if the node strength metric
should be used}

\item{sum_of_nearest_neighbors}{A logical value indicating if the sum of
nearest neighbors metric should be used}

\item{eigenvector_centrality}{A logical value indicating if the eigenvector
centrality metric should be used}
}
\value{
A named vector of weights for each metric
}
\description{
calculate weights for each metric
}
\examples{
# return weights for each metric
calculate_metrics_weight(betweenness_metric = TRUE, node_strength = TRUE,
                        sum_of_nearest_neighbors = TRUE, eigenvector_centrality = TRUE)
}
