% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{calculate_zero_raster}
\alias{calculate_zero_raster}
\title{Calculate raster objects for given extent and resolution
This function returns a list of zero raster and map grey background extent}
\usage{
calculate_zero_raster(
  geoscale,
  mean_index_raster,
  resolution = the$parameters_config$`CCRI parameters`$Resolution
)
}
\arguments{
\item{geoscale}{A list of longitude and latitude values for cropland analysis}

\item{mean_index_raster}{A raster object for mean index raster}

\item{resolution}{resolution to plot raster and map}
}
\value{
A list of zero raster and map grey background extent
}
\description{
Calculate raster objects for given extent and resolution
This function returns a list of zero raster and map grey background extent
}
