% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sean.R
\name{ccri}
\alias{ccri}
\title{Calculate Cropland Connectivity Risk Index (CCRI)}
\usage{
ccri(
  link_threshold = 0,
  power_law_metrics =
    the$parameters_config$`CCRI parameters`$NetworkMetrics$InversePowerLaw,
  negative_exponential_metrics =
    the$parameters_config$`CCRI parameters`$NetworkMetrics$NegativeExponential,
  rast,
  crop_cells_above_threshold,
  thresholded_crop_values
)
}
\arguments{
\item{link_threshold}{A threshold value for link}

\item{power_law_metrics}{A list of 2 vectors - power law metrics and weights}

\item{negative_exponential_metrics}{A list of of 2 vectors - negative exponential metrics}

\item{crop_cells_above_threshold}{A list of crop cells above threshold}

\item{thresholded_crop_values}{A list of crop values above threshold}
}
\value{
A list of calculated CCRI values
}
\description{
This function calculates CCRI for given parameters using power law and negative exponential.
It's required to call \code{\link[=initialize_cropland_data]{initialize_cropland_data()}} before calling this function.
}
\details{
Network metrics should be passed as a list of vectors e.g. \code{list(metrics = c("betweeness"), weights = c(100))}.
Default values are fetched from \code{parameters.yaml} and arguments uses the same structure.
}
\seealso{
\code{\link[=get_param_metrics]{get_param_metrics()}}, \code{\link[=sean]{sean()}}
}
