% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{ccri_diff}
\alias{ccri_diff}
\title{Calculate difference map}
\usage{
ccri_diff(rast, x, y, global, geoscale, resolution = reso())
}
\arguments{
\item{rast}{A raster object for mean index raster difference}

\item{x}{A raster object for cropland harvest}

\item{y}{A raster object for cropland harvest}

\item{geoscale}{geographical scale with longitude and latitudes. It will be converted into extent.
See \code{\link[=get_geographic_scales]{get_geographic_scales()}} and \code{\link[terra:ext]{terra::ext()}}.}

\item{resolution}{resolution to plot raster and map}
}
\description{
This function produces a map of difference b/w mean and sum indexes in rank of cropland harvested area fraction.
}
