% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{ccri_neg_exp_fun}
\alias{ccri_neg_exp_fun}
\title{Calculate CCRI using negative exponential for given parameters
This function calculates CCRI using negative exponential
for given parameters based on provided metrics and parameters.}
\usage{
ccri_neg_exp_fun(
  dispersal_parameter_gamma_val,
  link_threshold,
  distance_matrix = the$distance_matrix,
  thresholded_crop_values,
  crop_raster,
  crop_cells_above_threshold,
  metrics = the$parameters_config$`CCRI parameters`$NetworkMetrics$InversePowerLaw
)
}
\arguments{
\item{dispersal_parameter_gamma_val}{A list of gamma values}

\item{link_threshold}{A threshold value for link}

\item{distance_matrix}{distance matrix calculated during initialize_crop_data().}

\item{thresholded_crop_values}{crop values above threshold}

\item{crop_raster}{A raster object for crop raster}

\item{crop_cells_above_threshold}{A list of crop cells above threshold.
This is extracted in \code{\link[=initialize_cropland_data]{initialize_cropland_data()}}.}

\item{metrics}{A list 2 vectors - metrics and weights.}
}
\value{
A list of calculated CCRI values using negative exponential
}
\description{
Calculate CCRI using negative exponential for given parameters
This function calculates CCRI using negative exponential
for given parameters based on provided metrics and parameters.
}
