% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sean.R
\name{ccri_negative_exp}
\alias{ccri_negative_exp}
\title{Calculate negative exponential}
\usage{
ccri_negative_exp(
  dispersal_parameter_gamma_vals,
  link_threshold = 0,
  metrics = the$parameters_config$`CCRI parameters`$NetworkMetrics$InversePowerLaw,
  crop_cells_above_threshold = NULL,
  thresholded_crop_values = NULL
)
}
\arguments{
\item{dispersal_parameter_gamma_vals}{A list of gamma values}

\item{link_threshold}{A threshold value for link}

\item{metrics}{A list 2 vectors - metrics and weights.}

\item{crop_cells_above_threshold}{A list of crop cells above threshold}

\item{thresholded_crop_values}{A list of crop values above threshold}
}
\value{
A list of calculated negative exponential
}
\description{
Calculate negative exponential
}
