% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{ccri_negative_exponential}
\alias{ccri_negative_exponential}
\title{Calculate negative exponential}
\usage{
ccri_negative_exponential(
  dispersal_parameter_gamma_vals,
  link_threshold = 0,
  metrics = the$parameters_config$`CCRI parameters`$NetworkMetrics$InversePowerLaw,
  crop_cells_above_threshold = NULL,
  thresholded_crop_values = NULL
)
}
\arguments{
\item{dispersal_parameter_gamma_vals}{A list of gamma values}

\item{link_threshold}{A threshold value for link}

\item{crop_cells_above_threshold}{A list of crop cells above threshold}

\item{thresholded_crop_values}{A list of crop values above threshold}

\item{betweenness_metric}{A boolean value to calculate betweenness metric}

\item{node_strength}{A boolean value to calculate node strength}

\item{sum_of_nearest_neighbors}{A boolean value to calculate sum of nearest neighbors}

\item{eigenvector_centrality}{A boolean value to calculate eigenvector centrality}
}
\value{
A list of calculated negative exponential
}
\description{
Calculate negative exponential
}
