% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{ccri_powerlaw}
\alias{ccri_powerlaw}
\title{Calculate inverse power law}
\usage{
ccri_powerlaw(
  dispersal_parameter_beta_vals,
  link_threshold = 0,
  betweenness_metric = FALSE,
  node_strength = FALSE,
  sum_of_nearest_neighbors = FALSE,
  eigenvector_centrality = FALSE,
  crop_cells_above_threshold = NULL,
  thresholded_crop_values = NULL
)
}
\arguments{
\item{dispersal_parameter_beta_vals}{A list of beta values}

\item{link_threshold}{A threshold value for link}

\item{betweenness_metric}{A boolean value to calculate betweenness metric}

\item{node_strength}{A boolean value to calculate node strength}

\item{sum_of_nearest_neighbors}{A boolean value to calculate sum of nearest neighbors}

\item{eigenvector_centrality}{A boolean value to calculate eigenvector centrality}

\item{crop_cells_above_threshold}{A list of crop cells above threshold}

\item{thresholded_crop_values}{A list of crop values above threshold}
}
\value{
A list of calculated inverse power law
}
\description{
Calculate inverse power law
}
