% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{ccri_powerlaw_fun}
\alias{ccri_powerlaw_fun}
\title{Calculate CCRI using powerlaw for given parameters
This function calculates CCRI using powerlaw for given parameters based on provided metrics and parameters.}
\usage{
ccri_powerlaw_fun(
  dispersal_parameter_beta,
  link_threshold,
  distance_matrix = the$distance_matrix,
  thresholded_crop_values,
  crop_raster,
  crop_cells_above_threshold,
  metrics = the$parameters_config$`CCRI parameters`$NetworkMetrics$InversePowerLaw
)
}
\arguments{
\item{dispersal_parameter_beta}{A list of beta values}

\item{link_threshold}{A threshold value for link}

\item{distance_matrix}{distance matrix, generated during initialize_crop_data()}

\item{thresholded_crop_values}{crop values above threshold}

\item{crop_raster}{A raster object for cropland harvest}

\item{crop_cells_above_threshold}{crop cells above threshold. Only contains cells and not the the values.}

\item{metrics}{A list 2 vectors - metrics and weights.}
}
\value{
A list of calculated CCRI values using powerlaw
}
\description{
Calculate CCRI using powerlaw for given parameters
This function calculates CCRI using powerlaw for given parameters based on provided metrics and parameters.
}
