% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{ccri_powerlaw_function}
\alias{ccri_powerlaw_function}
\title{Calculate CCRI using powerlaw for given parameters
This function calculates CCRI using powerlaw for given parameters based on provided metrics and parameters.}
\usage{
ccri_powerlaw_function(
  dispersal_parameter_beta,
  link_threshold,
  distance_matrix = the$distance_matrix,
  thresholded_crop_values,
  crop_raster,
  crop_cells_above_threshold,
  betweenness_metric = FALSE,
  node_strength = FALSE,
  sum_of_nearest_neighbors = FALSE,
  eigenvector_centrality = FALSE
)
}
\arguments{
\item{dispersal_parameter_beta}{A list of beta values}

\item{link_threshold}{A threshold value for link}

\item{distance_matrix}{distance matrix, generated during initialize_crop_data()}

\item{thresholded_crop_values}{crop values above threshold}

\item{crop_raster}{A raster object for cropland harvest}

\item{crop_cells_above_threshold}{crop cells above threshold. Only contains cells and not the the values.}

\item{betweenness_metric}{A boolean value to calculate betweenness metric}

\item{node_strength}{A boolean value to calculate node strength}

\item{sum_of_nearest_neighbors}{A boolean value to calculate sum of nearest neighbors}

\item{eigenvector_centrality}{A boolean value to calculate eigenvector centrality}
}
\value{
A list of calculated CCRI values using powerlaw
}
\description{
Calculate CCRI using powerlaw for given parameters
This function calculates CCRI using powerlaw for given parameters based on provided metrics and parameters.
}
