% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{ccri_variance}
\alias{ccri_variance}
\title{Calculate variance of CCRI}
\usage{
ccri_variance(indexes, rast, global, geoscale, resolution = reso())
}
\arguments{
\item{indexes}{list of rasters}

\item{rast}{A raster object. It will be used in calculating variance.}

\item{geoscale}{vector. geographical scale}
}
\description{
This function produces a map of variance of CCRI based on input parameters
}
