% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccri_helper.R
\name{check_metrics}
\alias{check_metrics}
\title{Check if metrics in the list are valid}
\usage{
check_metrics(metrics_list)
}
\arguments{
\item{metrics_list}{A character vector of metrics to check}
}
\value{
A named logical vector indicating if each metric is valid or not
}
\description{
Check if metrics in the list are valid
}
\examples{
# return list of valid metrics
check_metrics(list("betweeness", "invalid_metric"))
}
