% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rasters.R
\name{cropharvest_rast}
\alias{cropharvest_rast}
\title{Get raster object for crop}
\usage{
cropharvest_rast(crop_name, data_source)
}
\arguments{
\item{crop_name}{Name of the crop}

\item{data_source}{Data source for cropland information}
}
\value{
Raster.
}
\description{
Get cropland information in a form of raster object from data source for crop
}
\examples{
\dontrun{
cropharvest_rast("avocado", "monfreda")
}

}
