% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{inv_powerlaw}
\alias{inv_powerlaw}
\alias{neg_exp}
\title{Dispersal kernels}
\usage{
inv_powerlaw(params = load_parameters())

neg_exp(params = load_parameters())
}
\arguments{
\item{params}{Object.\code{\link[=load_parameters]{load_parameters()}} by default.}
}
\value{
List with parameters and values.
}
\description{
-\verb{[inv_powerlaw()]} Get parameters and values pertaining to the inverse power law model.
-\verb{[neg_exp()]} Get parameters and values pertaining to the negative exponential model.
}
