% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hci_helper.R
\name{dist_methods}
\alias{dist_methods}
\title{Distance methods supported}
\usage{
dist_methods()
}
\value{
vector
}
\description{
Contains supported strategies to calculate distance between two points.
Use of one of two methods in \code{\link[=sean]{sean()}} or \code{\link[=sensitivity_analysis]{sensitivity_analysis()}}.
}
\examples{
dist_methods()

}
