% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccri_helper.R
\name{.check_yaml_structure}
\alias{.check_yaml_structure}
\title{Check structure of YAML file}
\usage{
.check_yaml_structure(existing_yaml_file, provided_yaml_file)
}
\arguments{
\item{existing_yaml_file}{Path to the existing YAML file}

\item{provided_yaml_file}{Path to the provided YAML file}
}
\value{
TRUE if the structures match, FALSE otherwise
}
\description{
Check the structure of the provided YAML file, and make sure it matches the structure of the existing YAML file.
}
