% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geohabnet-package.R
\docType{package}
\name{geohabnet-package}
\alias{geohabnet}
\alias{geohabnet-package}
\title{geohabnet: Analysis of cropland connectivity}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

c("Geographical spatial analysis of cropland connectivity. Allows users to visualize risk index plots for a given set of crops. Xing et al. (2021) \url{https://doi.org/10.1093/biosci/biaa067}. Package currently support crops sourced from monfreda and spam. The analysis produces 3 maps - mean, variance and difference for the risk index.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GarrettLab/CroplandConnectivity}
  \item Report bugs at \url{https://github.com/GarrettLab/CroplandConnectivity/issues}
}

}
\author{
\strong{Maintainer}: Krishna Keshav \email{kkeshav@ufl.edu}

Authors:
\itemize{
  \item Karen Garrett \email{karengarrett@ufl.edu} (\href{https://orcid.org/0000-0002-6578-1616}{ORCID})
}

Other contributors:
\itemize{
  \item Aaron Plex \email{plexaaron@ufl.edu} (\href{https://orcid.org/0000-0001-7317-3090}{ORCID}) [contributor]
  \item Garrett Lab [copyright holder, funder]
}

}
\keyword{internal}
