% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{GeoModel-class}
\alias{GeoModel-class}
\alias{.model_ob}
\title{GeoModel class}
\description{
A ref class to represent results of dispersal models.
}
\section{Fields}{

\describe{
\item{\code{matrix}}{An adjacency matrix to represent network.}
}}


