% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_network.R
\docType{class}
\name{GeoNetwork}
\alias{GeoNetwork}
\title{Class GeoNetwork.}
\description{
A class representing a network of geographical data.
This will wrap all the results from the risk analysis using \code{\link[=sean]{sean()}} or \code{\link[=sensitivity_analysis]{sensitivity_analysis()}}.
}
\section{Fields}{

\describe{
\item{\code{rasters}}{A list of \code{\link{GhabRasters}} objects.}
}}

