% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_rasters.R
\docType{class}
\name{GeoRasters}
\alias{GeoRasters}
\alias{.rast_ro}
\title{GeoRasters. Class to represent rasters from geohabnet}
\description{
A class to represent raster vis-a-vis risk indices.
This class encapsulates the results of apply dispersal models and metrics.
}
\section{Fields}{

\describe{
\item{\code{rasters}}{List. List of raster representing risk indices.}

\item{\code{global}}{Boolean. True if contains \code{GlobalRast} object, False otherwise.}
}}


