% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{geoscale_param}
\alias{geoscale_param}
\title{Get geographical scales from the parameters}
\usage{
geoscale_param(gparams = load_parameters())
}
\arguments{
\item{gparams}{Optional. \code{\link[=load_parameters]{load_parameters()}} or null}
}
\value{
Vector. A set of geographical scales
}
\description{
This function returns a list of geographical scales set in global and custom extent in \code{parameters.yaml}.
If \code{global} is \code{TRUE}, the \code{CustomExt} is ignored.
}
