% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{get_cropharvest_raster_sum}
\alias{get_cropharvest_raster_sum}
\title{Get sum of rasters for individual crops}
\usage{
get_cropharvest_raster_sum(crop_names)
}
\arguments{
\item{crop_names}{A named list of source along with crop names}
}
\value{
Raster object which is sum of all the individual crop rasters
}
\description{
Takes crop names and returns raster object which is sum of raster of individual crops.
Currently, only supports crops listed in
\code{\link[geodata:crop_monfreda]{geodata::monfredaCrops()}}, \code{\link[geodata:crop_spam]{geodata::spamCrops()}}
If crop is present in multiple sources, then their mean is calculated.
}
\examples{
\dontrun{
get_cropharvest_raster_sum(list(monfreda = c("wheat", "barley"), spam = c("wheat", "potato")))
}
}
