% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{get_geographic_scales}
\alias{get_geographic_scales}
\title{Get geographical scales from the paramters
This function returns a list of geographical scales set global and custom extent in parameters.yaml}
\usage{
get_geographic_scales()
}
\value{
A list of geographical scales
}
\description{
Get geographical scales from the paramters
This function returns a list of geographical scales set global and custom extent in parameters.yaml
}
