% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_param_metrics}
\alias{get_param_metrics}
\title{Get metrics from parameters}
\usage{
get_param_metrics(params = load_parameters())
}
\arguments{
\item{params}{R object of \code{\link[=load_parameters]{load_parameters()}}. Default is \code{load_parameters()}.}
}
\value{
List. List of metrics - parameters and values. See usage.
}
\description{
Get metrics and parameters stored in \code{parameters.yaml}.
}
\examples{
# Get metrics from parameters
get_param_metrics()
get_param_metrics(load_parameters())

}
