% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rasters.R
\name{get_rasters}
\alias{get_rasters}
\title{Get rasters object from parameters}
\usage{
get_rasters(hosts)
}
\arguments{
\item{hosts}{List of hosts and values. It is synonym to Hosts object in parameters}
}
\value{
List of rasters
}
\description{
Takes named list of hosts as an input. See host object in \code{\link[=get_parameters]{get_parameters()}} or \code{\link[=load_parameters]{load_parameters()}}.
Function creates 2 raster object - one is a sum of all the crops specified under sources
and other using the provided raster file. See \code{\link[=get_crop_raster_fromtif]{get_crop_raster_fromtif()}}
}
\examples{
# Get default rasters
\dontrun{
get_rasters(list(spam = c("wheat"), monfreda = c("avocado"), file = "some_raster.tif"))
}
}
\seealso{
\code{\link[=load_parameters]{load_parameters()}}, \code{\link[=get_parameters]{get_parameters()}}, \code{\link[=get_crop_raster_fromtif]{get_crop_raster_fromtif()}}, \code{\link[=get_cropharvest_raster]{get_cropharvest_raster()}}
}
