% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccri_helper.R
\name{get_supported_sources}
\alias{get_supported_sources}
\title{Get supported sources of crops}
\usage{
get_supported_sources()
}
\value{
Vector of supported sources.
Also used as a lookup to find get raster object.
}
\description{
When provided, \code{\link[=cropharvest_rast]{cropharvest_rast()}} will
look for cropland data in this specific source.
}
\examples{
# Get currently supported sources
get_supported_sources()
}
