% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{global_analysis}
\alias{global_analysis}
\title{Global cropland density map
Only when user has enabled global analysis}
\usage{
global_analysis(
  map_grey_background_extent,
  resolution = the$parameters_config$`CCRI parameters`$Resolution
)
}
\arguments{
\item{map_grey_background_extent}{A raster object for map's grey background}

\item{resolution}{resolution to plot raster and map}
}
\description{
Global cropland density map
Only when user has enabled global analysis
}
