% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{global_scales<-}
\alias{global_scales<-}
\title{Set global geographical extent}
\usage{
global_scales(east, west) <- value
}
\arguments{
\item{east}{vector. The eastern hemisphere extent.}

\item{west}{vector. The western hemisphere extent.}
}
\description{
Set the geographical extents used in global analysis.
Each extent should be in the form of c(Xmin, Xmax, Ymin, Ymax)
}
\seealso{
\code{\link[=global_scales]{global_scales()}}
\code{\link[terra:ext]{terra::ext()}}
}
