% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_rasters.R
\docType{class}
\name{GlobalRast-class}
\alias{GlobalRast-class}
\alias{.grast_ro}
\title{GlobalRast class}
\description{
A class to represent raster objects for global scales. Global scales are accessible using \code{\link[=global_scales]{global_scales()}}.
However, this class encapsulates the results of apply dispersal models and metrics.
}
\section{Fields}{

\describe{
\item{\code{east}}{A list of raster objects for eastern hemisphere.}

\item{\code{west}}{A list of raster objects for western hemisphere.}
}}


