% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_network.R
\docType{class}
\name{Gmap}
\alias{Gmap}
\title{Maps}
\description{
A class to represent various maps.
}
\section{Fields}{

\describe{
\item{\code{me_rast}}{SpatRaster A raster representing mean risk index.}

\item{\code{me_out}}{Character. A file path to the mean risk index raster.}

\item{\code{diff_rast}}{SpatRaster A raster representing difference.}

\item{\code{diff_out}}{Character. A file path to the difference raster.}

\item{\code{var_rast}}{Numeric. A raster representing variance.}

\item{\code{var_out}}{SpatRaster A file path to the variance raster.}
}}

