% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_network.R
\name{habitat_density<-}
\alias{habitat_density<-}
\title{Set the habitat density.}
\usage{
habitat_density(x) <- value
}
\arguments{
\item{x}{the GeoNetwork object.}

\item{value}{SpatRaster.}
}
\value{
The same object type as x, that is, GeoNetwork. This function returns the updated S4 GeoNetwork object with the new habitat availability SpatRaster assigned.
}
\description{
This function helps to set a \code{SpatRaster} of the habitat availability or density in a GeoNetwork object. The function is an S4 replacement method in the geohabnet package. It allows you to assign a host availability \code{SpatRaster} to a geohabnet object.
}
