% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_network.R
\name{host_density<-}
\alias{host_density<-}
\title{Set the host density.}
\usage{
host_density(x) <- value
}
\arguments{
\item{x}{the GeoNetwork object.}

\item{value}{SpatRaster.}
}
\description{
Sets the host density slot in the GeoNetwork object
}
