% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{initialize_cropland_data}
\alias{initialize_cropland_data}
\title{intialize cropland data with geiven paramters, it will be later used to calculate CCRI and other functions}
\usage{
initialize_cropland_data(
  cropharvest_raster,
  resolution = 12,
  geo_scale,
  host_density_threshold = 0,
  agg_method = "sum"
)
}
\arguments{
\item{cropharvest_raster}{A raster object for cropland harvest}

\item{resolution}{resolution to plot raster and map  (default: 12)}

\item{geo_scale}{A list of longitude and latitude values for cropland analysis}

\item{host_density_threshold}{A threshold value for cropland density (default: 0)}

\item{agg_method}{A method to aggregate cropland raster (default: "sum")}
}
\description{
intialize cropland data with geiven paramters, it will be later used to calculate CCRI and other functions
}
\details{
This function also creates global variables which are result of applying aggregate functions into raster.
Theese global variables are used when applying allgorithms - ipl\code{\link[=ccri_powerlaw]{ccri_powerlaw()}} and ne\code{\link[=ccri_negative_exp]{ccri_negative_exp()}}.
}
