% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{load_parameters}
\alias{load_parameters}
\title{Load Parameters from YAML File}
\usage{
load_parameters(filepath = .param_fp())
}
\arguments{
\item{filepath}{Path to the YAML file containing the parameters. By default, it
takes the value of \code{parameters.yaml} in R user's directory.}
}
\value{
object with parameters and values
}
\description{
This function loads parameters from a YAML file and stores them in an object.
}
\examples{
# Load parameters from default file
load_parameters()

}
