% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_metrics.R
\docType{data}
\name{metric_funs}
\alias{metric_funs}
\title{list stores functions to apply metrics to distance matrix}
\format{
An object of class \code{list} of length 7.
}
\usage{
metric_funs
}
\description{
list stores functions to apply metrics to distance matrix
}
\keyword{internal}
