% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{Models}
\alias{Models}
\alias{.model_ob}
\title{Models}
\description{
A class to represent results of dispersal models.
}
\section{Fields}{

\describe{
\item{\code{matrix}}{An adjacency matrix to represent network.}
}}


