% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{plot_maps}
\alias{plot_maps}
\title{Calculate and plot maps
Calculate mean, variance and difference. The result is produced in form of maps plotted with predefined settings.
Currently, the settings for plot cannot be customized.}
\usage{
plot_maps(indexes, geoscale, reso, pmean = TRUE, pvar = TRUE, pdiff = TRUE)
}
\arguments{
\item{indexes}{list of raster}

\item{geoscale}{geographical scale}

\item{reso}{map resolution}

\item{pmean}{\code{TRUE} if map of mean should be plotted, \code{FALSE} otherwise}

\item{pvar}{\code{TRUE} if variance map should be plotted, \code{FALSE} otherwise}

\item{pdiff}{\code{TRUE} if difference map should be plotted, \code{FALSE} otherwise}
}
\description{
Calculate and plot maps
Calculate mean, variance and difference. The result is produced in form of maps plotted with predefined settings.
Currently, the settings for plot cannot be customized.
}
