% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{reset_params}
\alias{reset_params}
\title{Reset parameters.yaml}
\usage{
reset_params()
}
\description{
Resets the values in the \code{parameters.yaml}
file to the default initial values.
}
\examples{
\dontrun{
reset_params()
}
}
