% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{reso<-}
\alias{reso<-}
\title{Set resolution value}
\usage{
reso(reso) <- value
}
\arguments{
\item{reso}{resolution value. Must be greater than 0 and less than or equal to 48.}
}
\description{
Set \code{resolution} to be used in analysis.
It doesn't modify the \code{parameters.yaml}
}
