% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{reso}
\alias{reso}
\title{Get resolution value}
\usage{
reso()
}
\value{
Numeric. Resolution from \code{parameters.yaml}. Default is 24.
}
\description{
Resolution stored in \code{parameter.yaml}.
If not present it will result default value.
}
