% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hci_helper.R
\name{risk_indices}
\alias{risk_indices}
\title{Get risk indices}
\usage{
risk_indices(ri)
}
\arguments{
\item{ri}{GeoRasters object}
}
\value{
List of risk indices. If the \code{ri} is global, the list will contain two elements,
one for each hemisphere. e.g. \code{list(east = list(), west = list())}. If the \code{ri} is not global,
the list will contain a single element, e.g. \code{list()}.
}
\description{
Get risk indices from GeoRasters object.
}
\details{
This function will unpack SpatRasters from GeoModel and thus is \code{\link[future:future]{future::future()}} safe.
}
