% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskmap.R
\docType{class}
\name{RiskMap-class}
\alias{RiskMap-class}
\alias{RiskMap}
\title{RiskMap class}
\description{
An S4 class representing resulting maps from specific operation types.
Handles automated defaults for missing or NULL spatial data.
}
