% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskmap.R
\docType{class}
\name{RiskMap}
\alias{RiskMap}
\title{RiskMap. Class to represent risk maps from geohabnet}
\description{
A class representing resulting maps from the specific operation type.
}
\section{Fields}{

\describe{
\item{\code{map}}{Character. A file path to the map.}

\item{\code{riid}}{SpatRaster. This is one of the risk maps.}

\item{\code{spr}}{SpatRaster. A spatial raster representing the risk index.}

\item{\code{fp}}{Character. A file path to the risk index raster.}
}}

