% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sean.R
\name{sa_onrasters}
\alias{sa_onrasters}
\alias{msean_onrast}
\title{Run sensitivity analysis}
\usage{
sa_onrasters(rast, link_thresholds = c(0), hd_thresholds = c(0), ...)

msean_onrast(
  global = TRUE,
  geoscale = NULL,
  res = reso(),
  outdir = tempdir(),
  ...
)
}
\arguments{
\item{rast}{This is a SpatRaster object that indicates habitat suitability across locations in a landscape. Entry values in this object can include any positive real numbers, but they typically range from 0 to 1. Currently, this argument supports only raster layers in the standard global coordinate reference system -CRS-, that is, WGS 84 or World Geodetic System 1984, identified by the EPSG code 4326. Users need to project the raster layer before providing it as an input in \code{\link[=msean]{msean()}} if it is in another CRS.}

\item{link_thresholds}{Numeric vector. Threshold values for link weight.
Based on the information on the habitat layer and dispersal kernels, adjacency matrices are created, where entries represent the potential of an organism's movement between habitat locations.
Then, the adjacency matrices are converted into graph objects to perform a network analysis, where the entries in the adjacency matrices are now the werights of the links of the network.
This parameter supports any positive values, but make sure these values are smaller than the maximum link weight in the network.
If \code{link_threshold = 0}, all links int he netowkr will be considered in the connectivity analysis.
Choosing link weight thresholds greater than zero helps to focus the analysis on the msot likely dispersal connections in the landscape.
All link weights that are below this threshold will be replaced with zeros for the connectivity analysis.}

\item{hd_thresholds}{Numeric vector. host density threshold values}

\item{...}{Additional parameters to be passed to \code{\link[=sean]{sean()}}.}

\item{global}{Logical. \code{TRUE} if global analysis, \code{FALSE} otherwise.
Default is \code{TRUE}}

\item{geoscale}{Numeric vector. This is a set of geographic coordinates in the form of
c(Xmin, Xmax, Ymin, Ymax) which correspond to the positions in the standard geographic coordinate system or EPSG:4326. \code{geoscape} referst to the geographic extent, or bounding box, for the habitat connectivity analysis.
Habitat locations outside the geographic extent will be excluded from the analysis.
If \code{geoscale} is NuLL, the extent is extracted from \code{rast}(SpatRaster) using \code{\link[terra:ext]{terra::ext()}}.}

\item{res}{Numeric. The spatial aggregation factor that will be used to aggregate the raster layer of habitat availability, from fine to coars resolution. Default is \code{\link[=reso]{reso()}}.}

\item{outdir}{Character. Output directory for saving raster in TIFF format.
Default is \code{\link[=tempdir]{tempdir()}}.}
}
\value{
A list of calculated CCRI indices after operations.
An index is generated for each combination of paramters.
One combination is equivalent to \code{\link[=sean]{sean()}} function.
}
\description{
Same as \code{\link[=sensitivity_analysis]{sensitivity_analysis()}} but it takes raster object and other parameters as an input.
\itemize{
\item \code{sa_onrasters()} is a wrapper around \code{\link[=sean]{sean()}} function. Takes raster object and other parameters as an input.
\item \code{msean_onrast()} same as \code{\link[=sa_onrasters]{sa_onrasters()}}. Use this for side effects + results.
Produces and plots the maps for the outcomes and results are returned as an object.
It produces and plots the maps for the outcomes and results are returned as an object.
}
}
\details{
Error not handled for non-overlapping extents.
}
\references{
Yanru Xing, John F Hernandez Nopsa, Kelsey F Andersen, Jorge L Andrade-Piedra, Fenton D Beed,
Guy Blomme, Mónica Carvajal-Yepes, Danny L Coyne, Wilmer J Cuellar, Gregory A Forbes,
Jan F Kreuze, Jürgen Kroschel, P Lava Kumar, James P Legg, Monica Parker, Elmar Schulte-Geldermann,
Kalpana Sharma, Karen A Garrett,
\emph{Global Cropland connectivity: A Risk Factor for Invasion and Saturation by Emerging Pathogens and Pests},
BioScience, Volume 70, Issue 9, September 2020, Pages 744–758,
\doi{10.1093/biosci/biaa067}

Hijmans R (2023). \emph{terra: Spatial Data Analysis}.
R package version 1.7-46, \url{https://CRAN.R-project.org/package=terra}
}
\seealso{
Use \code{\link[=get_rasters]{get_rasters()}} to obtain raster object.

\code{\link[=msean_onrast]{msean_onrast()}}
}
