% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{scales}
\alias{scales}
\title{Only meant to global variables}
\format{
An object of class \code{environment} of length 2.
}
\usage{
scales
}
\description{
Only meant to global variables
}
\keyword{internal}
