% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sean.R
\name{sean_linkweights}
\alias{sean_linkweights}
\title{Calculate sensitivity analysis on cropland harvested area fraction}
\usage{
sean_linkweights(
  link_threshold = 0,
  host_density_thresholds,
  global = TRUE,
  geo_scale,
  aggregate_methods,
  rast,
  resolution,
  maps = TRUE
)
}
\arguments{
\item{link_threshold}{A threshold value for link}

\item{host_density_thresholds}{A list of host density threshold values}

\item{global}{\code{TRUE} if global analysis, \code{FALSE} otherwise. Default is \code{TRUE}}

\item{geo_scale}{A list of longitude and latitude values for cropland analysis}

\item{aggregate_methods}{A list of aggregation methods}

\item{resolution}{resolution to plot raster and map}

\item{cropharvest_raster}{A raster object for cropland harvest}
}
\value{
A list of calculated CCRI values using negative exponential
}
\description{
This function calculates sensitivity analysis on cropland harvested area fraction based on link weight threshold
and other provided parameters.
It can be used as entry point for sensitivity analysis.
}
\seealso{
\code{\link[=sa_onrasters]{sa_onrasters()}}
\link{sean_linkweights}
\code{\link[=sean]{sean()}}
\code{\link[=plot_maps]{plot_maps()}}
}
