% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccri_helper.R
\name{search_crop}
\alias{search_crop}
\title{Search for crop}
\usage{
search_crop(name)
}
\arguments{
\item{name}{name of crop}
}
\description{
It returns the dataset sources in which crop data is available.
It's a wrapper around \code{\link[geodata:crop_spam]{geodata::spamCrops()}} and \code{\link[geodata:crop_monfreda]{geodata::monfredaCrops()}}
}
\examples{
search_crop("coffee")
search_crop("wheat")
\donttest{
search_crop("jackfruit")
}

}
\seealso{
\code{\link[=get_supported_sources]{get_supported_sources()}}
}
