% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sean.R
\name{sensitivity_analysis}
\alias{sensitivity_analysis}
\title{Calculate sensitivity analysis on parameters}
\usage{
sensitivity_analysis()
}
\description{
This function runs sensitivity analysis on parameters based on provided parameters through \code{\link[=set_parameters]{set_parameters()}}.
It can be used as entry point for sensitivity analysis.
Plots results of sensitivity analysis.
}
\examples{
\dontrun{
# Run analysis on specified parameters.yaml
sensitivity_analysis()
}
}
\seealso{
\code{\link[=sa_onrasters]{sa_onrasters()}}
\link{sean_linkweights}
\code{\link[=sean]{sean()}}
\code{\link[=plot_maps]{plot_maps()}}
}
