% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{sensitivity_analysis_on_geoscale}
\alias{sensitivity_analysis_on_geoscale}
\title{Calculate sensitivity analysis on cropland harvested area fraction}
\usage{
sensitivity_analysis_on_geoscale(
  link_thresholds,
  host_density_thresholds,
  geo_scale,
  aggregate_methods = c("sum", "mean"),
  cropharvest_raster,
  resolution
)
}
\arguments{
\item{link_thresholds}{A list of threshold values for link}

\item{host_density_thresholds}{A list of host density threshold values}

\item{geo_scale}{longitude and latitude values for cropland analysis}

\item{aggregate_methods}{A list of aggregation methods}

\item{cropharvest_raster}{A raster object for cropland harvest}

\item{resolution}{resolution to plot raster and map}
}
\value{
A list of calculated CCRI values using negative exponential
}
\description{
This function calculates sensitivity analysis on cropland harvested area fraction
based on geographical scale threshold and other provided parameters.
It can be used as entry point for sensitivity analysis.
}
