% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{sensitivity_analysis_on_link_weight}
\alias{sensitivity_analysis_on_link_weight}
\title{Calculate sensitivity analysis on cropland harvested area fraction
This function calculates sensitivity analysis on cropland harvested area fraction based on provided parameters.
It can be used as entry point for sensitivity analysis.}
\usage{
sensitivity_analysis_on_link_weight(
  link_threshold = 0,
  host_density_thresholds,
  geo_scale,
  aggregate_methods,
  cropharvest_raster,
  resolution
)
}
\arguments{
\item{link_threshold}{A threshold value for link}

\item{host_density_thresholds}{A list of host density threshold values}

\item{geo_scale}{A list of longitude and latitude values for cropland analysis}

\item{aggregate_methods}{A list of aggregation methods}

\item{cropharvest_raster}{A raster object for cropland harvest}

\item{resolution}{resolution to plot raster and map}
}
\value{
A list of calculated CCRI values using negative exponential
}
\description{
Calculate sensitivity analysis on cropland harvested area fraction
This function calculates sensitivity analysis on cropland harvested area fraction based on provided parameters.
It can be used as entry point for sensitivity analysis.
}
