% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{sensitivity_analysis_on_params}
\alias{sensitivity_analysis_on_params}
\title{Calculate sensitivity analysis on cropland harvested area fraction}
\usage{
sensitivity_analysis_on_params(
  link_threshold = 0,
  geo_scale,
  aggregate_methods = c("sum", "mean"),
  cropharvest_raster,
  host_density_threshold = 0,
  resolution = 24
)
}
\arguments{
\item{link_threshold}{A threshold value for link}

\item{geo_scale}{A list of longitude and latitude values for cropland analysis}

\item{aggregate_methods}{A list of aggregation methods. It can be sum or mean.}

\item{cropharvest_raster}{A raster object for cropland harvest}

\item{host_density_threshold}{A host density threshold value}

\item{resolution}{resolution to plot raster and map}
}
\value{
A list of calculated CCRI values using negative exponential
}
\description{
This function calculates sensitivity analysis on cropland harvested area fraction based on provided parameters.
It can be used as entry point for sensitivity analysis.
}
\seealso{
\code{\link[=plot_maps]{plot_maps()}}
}
