% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{senstivity_analysis}
\alias{senstivity_analysis}
\title{Calculate sensitivity analysis on parameters
This function runs sensitivity analysis on parameters based on provided parameters thorugh set_parameters() function.
It can be used as entry point for sensitivity analysis.
Plots results of sensitivity analysis.}
\usage{
senstivity_analysis()
}
\description{
Calculate sensitivity analysis on parameters
This function runs sensitivity analysis on parameters based on provided parameters thorugh set_parameters() function.
It can be used as entry point for sensitivity analysis.
Plots results of sensitivity analysis.
}
