% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{set_global_scales}
\alias{set_global_scales}
\title{Set global geographical extent}
\usage{
set_global_scales(value)
}
\arguments{
\item{value}{list. Named list of extents for eastern and western hemisphere. See usage.}
}
\description{
Set the geographical extents used in global analysis.
Each extent should be in the form of c(Xmin, Xmax, Ymin, Ymax)
}
\examples{
\dontrun{
set_global_scales(list(east = c(Xmin = -24, Xmax = 180, Ymin = -58, Ymax = 60),
west = c(Xmin = -140, Xmax = -34, Ymin = -58, Ymax = 60)))
}
}
\seealso{
\code{\link[=global_scales]{global_scales()}}
\code{\link[terra:ext]{terra::ext()}}
}
