% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{set_parameters}
\alias{set_parameters}
\title{Set Parameters}
\usage{
set_parameters(new_params, iwindow = FALSE)
}
\arguments{
\item{new_params}{The path to the new parameters file.}

\item{iwindow}{Logical, indicating whether to prompt the user to select the
new parameters file using a file selection window. The default value of this parameter is set to FALSE.}
}
\value{
None
}
\description{
This function allows the user to set the parameters by replacing the existing
parameters file with a new one. Use \code{\link[=get_parameters]{get_parameters()}} to modify the parameter values.
}
\examples{
param_fp <- get_parameters()
set_parameters(param_fp)

}
