% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{set_parameters_object}
\alias{set_parameters_object}
\title{Set Parameters function}
\usage{
set_parameters_object(
  dispersal_parameter_beta = c(0.5, 1, 1.5),
  dispersal_parameter_gamma = c(0.05, 1, 0.2, 0.3),
  aggregation_strategy = c("sum", "mean"),
  hosts = c("avocado"),
  host_density_threshold = c(0.0015, 0.002, 0.0025),
  link_threshold = c(0, 1e-06, 6e-04),
  resolution = 24,
  global_analysis = FALSE,
  west_extent = c(-24, -180, -58, 60),
  east_extent = c(-140, -34, --58, 60),
  custom_extent = list(c(-115, -75, -5, 32)),
  metrics_inv_powerlaw = c("betweeness", "node_strength", "sum_of_nearest_neighbors",
    "eigenvector_centrality"),
  metrics_neg_exponential = c("betweeness", "node_strength", "sum_of_nearest_neighbors",
    "eigenvector_centrality")
)
}
\arguments{
\item{dispersal_parameter_beta}{Numeric vector of dispersal parameter beta values}

\item{dispersal_parameter_gamma}{Numeric vector of dispersal parameter gamma values}

\item{aggregation_strategy}{Character vector of aggregation strategies}

\item{hosts}{Character vector of hosts}

\item{host_density_threshold}{Numeric vector of host density threshold values}

\item{link_threshold}{Numeric vector of link threshold values}

\item{resolution}{Numeric vector of resolution values}

\item{global_analysis}{Logical vector of global analysis values}

\item{west_extent}{Numeric vector of west extent values}

\item{east_extent}{Numeric vector of east extent values}

\item{custom_extent}{List of custom extent values}

\item{metrics_inv_powerlaw}{Character vector of inv_powerlaw metrics}

\item{metrics_neg_exponential}{Character vector of neg_exponential metrics}
}
\value{
TRUE if the parameters were set successfully, FALSE otherwise
}
\description{
This function allows you to override existing parameters with new values.
}
\examples{
\dontrun{
# Set parameters
set_parameters_object()
# Set parameters with custom beta values
set_parameters_object(dispersal_parameter_beta = c(0.5, 1, 1.5))
}
}
\seealso{
\code{\link[=load_parameters]{load_parameters()}} \code{\link[=set_parameters]{set_parameters()}}
}
