% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{set_reso}
\alias{set_reso}
\title{Set resolution value}
\usage{
set_reso(value)
}
\arguments{
\item{value}{numeric. Resolution value.}
}
\value{
Invisible TRUE
}
\description{
Set \code{resolution} to be used in analysis.
It doesn't modify the \code{parameters.yaml}
but instead a currently loaded instance of it.
Must be greater than 0 and less than or equal to 48.
}
\examples{
set_reso(24)

}
