% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{setprops,GeoModel,character,numeric,numeric-method}
\alias{setprops,GeoModel,character,numeric,numeric-method}
\title{Set properties of the GeoModel object.}
\usage{
\S4method{setprops}{GeoModel,character,numeric,numeric}(x, aggregation, hdthreshold, linkthreshold)
}
\arguments{
\item{x}{The GeoModel object.}

\item{aggregation}{Character. A value representing the aggregation method.}

\item{hdthreshold}{Numeric. A value representing the host density threshold.}

\item{linkthreshold}{Numeric. A value representing the link threshold in a network.}
}
\value{
The GeoModel object with updated properties.
}
\description{
Set properties of the GeoModel object.
}
