% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_metrics.R
\name{sonn}
\alias{sonn}
\alias{node_strength}
\alias{betweeness}
\alias{ev}
\title{Calculation on network matrix.
These are basically an abstraction of functions under the \link{igraph} package.}
\usage{
sonn(crop_dm, we)

node_strength(crop_dm, we)

betweeness(crop_dm, we)

ev(crop_dm, we)
}
\arguments{
\item{crop_dm}{Distance matrix.
In the internal workflow, the distance matrix comes from \code{\link[=initialize_cropland_data]{initialize_cropland_data()}} and risk functions.}

\item{we}{Weight in percentage.}
}
\value{
Matrix with the mean value based on the assigned weight.
}
\description{
The functions included in this abstraction are:
\itemize{
\item \code{sonn()}: Calculates the sum of nearest neighbors.
\item \code{node_strength()}: Calculates the sum of edge weights of adjacent nodes.
\item \code{betweeness()}: Calculates the vertex and edge betweenness based on the number of geodesics.
\item \code{ev()}: Calculates the eigenvector centralities of positions within the network.
}
}
