% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\docType{data}
\name{STR_NODE_STRENGTH}
\alias{STR_NODE_STRENGTH}
\title{global variable containing string. Not intended to modify.}
\format{
An object of class \code{character} of length 1.
}
\usage{
STR_NODE_STRENGTH
}
\description{
global variable containing string. Not intended to modify.
}
\keyword{internal}
