% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_metrics.R
\name{supported_metrics}
\alias{supported_metrics}
\title{Returns metrics currently supported in the analysis.}
\usage{
supported_metrics()
}
\value{
vector of supported metrics.
}
\description{
Returns metrics currently supported in the analysis.
}
\examples{
supported_metrics()

}
