% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccri_helper.R, R/hci_helper.R
\name{supported_sources}
\alias{supported_sources}
\title{Get supported sources of crops}
\usage{
supported_sources()

supported_sources()
}
\value{
Vector of supported sources.
Also used as a lookup to find get raster object.

Vector of supported sources.
Also used as a lookup to find get raster object.
}
\description{
When provided, \code{\link[=cropharvest_rast]{cropharvest_rast()}} will
look for cropland data in this specific source.

When provided, \code{\link[=cropharvest_rast]{cropharvest_rast()}} will
look for cropland data in this specific source.
}
\examples{
# Get currently supported sources
supported_sources()
# Get currently supported sources
supported_sources()
}
