% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\docType{data}
\name{the}
\alias{the}
\title{Only meant to global variables}
\format{
An object of class \code{environment} of length 7.
}
\usage{
the
}
\description{
Only meant to global variables
}
\keyword{internal}
