% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{calculate_difference_map}
\alias{calculate_difference_map}
\title{Calculate difference map
This function produces a map of difference b/w mean and sum indexes in rank of cropland harvested area fraction.}
\usage{
calculate_difference_map(
  mean_index_rast,
  cropharvest_aggtm_crop,
  cropharvest_agglm_crop,
  geoscale,
  resolution = the$parameters_config$`CCRI parameters`$Resolution
)
}
\arguments{
\item{cropharvest_aggtm_crop}{A raster object for cropland harvest}

\item{cropharvest_agglm_crop}{A raster object for cropland harvest}

\item{resolution}{resolution to plot raster and map}

\item{mean_index_raster_diff}{A raster object for mean index raster difference}
}
\description{
Calculate difference map
This function produces a map of difference b/w mean and sum indexes in rank of cropland harvested area fraction.
}
