% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{ccri_diff}
\alias{ccri_diff}
\title{Calculate difference map}
\usage{
ccri_diff(rast, x, y, global, geoscale, res = reso())
}
\arguments{
\item{rast}{A raster object for mean index raster difference}

\item{x}{A raster object for cropland harvest}

\item{y}{A raster object for cropland harvest}

\item{global}{logical. \code{TRUE} if global analysis is required, \code{FALSE} otherwise.
When \code{TRUE}, \code{geoscale} is ignored. Default is \code{TRUE}.}

\item{geoscale}{vector. geographical scale}

\item{res}{numeric. map resolution}
}
\description{
This function produces a map of difference b/w mean and sum indexes in rank of cropland harvested area fraction.
}
