% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{ccri_diff}
\alias{ccri_diff}
\title{Calculate difference map}
\usage{
ccri_diff(rast, x, y, global, geoscale, res = reso(), outdir = tempdir())
}
\arguments{
\item{rast}{A raster object for mean index raster difference}

\item{x}{A raster object for cropland harvest}

\item{y}{A raster object for cropland harvest}

\item{global}{logical. \code{TRUE} if global analysis is required, \code{FALSE} otherwise.
When \code{TRUE}, \code{geoscale} is ignored. Default is \code{TRUE}.}

\item{geoscale}{Vector. geographical scale. Default is \code{NULL}.}

\item{res}{numeric. map resolution.}

\item{outdir}{Character. Output directory for saving raster in TIFF format.
Default is \code{\link[=tempdir]{tempdir()}}.}
}
\value{
RiskMap. \link{?RiskMap}. It contains result in the form of \code{SpatRaster} object
and filenames of the saved maps.
}
\description{
This function produces a map of difference b/w mean and sum indexes in rank of cropland harvested area fraction.
}
