% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{ccri_diff}
\alias{ccri_diff}
\title{Calculate difference map}
\usage{
ccri_diff(x, y, geoscale, res = reso(), outdir = tempdir())
}
\arguments{
\item{x}{SpatRaster.}

\item{y}{SpatRaster.}

\item{geoscale}{Numeric vector. \code{x} will be cropped to this extent.}

\item{res}{Numeric. map resolution.}

\item{outdir}{Character. Output directory for saving raster in TIFF format.
Default is \code{\link[=tempdir]{tempdir()}}.}

\item{rast}{SpatRaster. A template raster to hold the cell-wise difference}
}
\value{
RiskMap. Contains result in the form of \code{SpatRaster} objects
and file path of the saved maps.
}
\description{
This function produces a map of difference b/w mean and sum indexes in rank of cropland harvested area fraction.
}
