% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{ccri_mean}
\alias{ccri_mean}
\title{Calculate mean of raster objects}
\usage{
ccri_mean(indexes, global = TRUE, geoscale = NULL, reso = reso(), plt = TRUE)
}
\arguments{
\item{indexes}{list of rasters}

\item{geoscale}{vector. geographical scale}

\item{reso}{numeric, map resolution}

\item{plt}{\code{TRUE} if need to plot mean map, \code{FALSE} otherwise and \code{geoscale}, \code{reso} is ignored.}
}
\description{
Overriding for \code{\link[terra:summarize-generics]{terra::mean()}}. Calculates mean of list of rasters.
}
