% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{ccri_mean}
\alias{ccri_mean}
\title{Calculate mean of raster objects}
\usage{
ccri_mean(
  indexes,
  global = TRUE,
  geoscale = NULL,
  plt = TRUE,
  outdir = tempdir()
)
}
\arguments{
\item{indexes}{list of rasters. See details.}

\item{global}{logical. \code{TRUE} if global analysis is required, \code{FALSE} otherwise.
When \code{TRUE}, \code{geoscale} is ignored. Default is \code{TRUE}.}

\item{geoscale}{vector. geographical scale}

\item{plt}{\code{TRUE} if need to plot mean map, \code{FALSE} otherwise and \code{geoscale}.}

\item{outdir}{Character. Output directory for saving raster in TIFF format.
Default is \code{\link[=tempdir]{tempdir()}}.}
}
\description{
Wrapper for \code{\link[terra:summarize-generics]{terra::mean()}}. Calculates mean of list of rasters.
}
