% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{ccri_mean}
\alias{ccri_mean}
\title{Calculate mean of raster objects
Overriding for \code{\link[terra:summarize-generics]{terra::mean()}}. Calculates mean of list of rasters.}
\usage{
ccri_mean(indexes, geoscale = NULL, reso = NULL, plt = TRUE)
}
\arguments{
\item{indexes}{raster list}

\item{geoscale}{geographical scale for plotting}

\item{reso}{resolution for plotting}

\item{plt}{\code{TRUE} if need to plot mean map, \code{FALSE} otherwise and \code{geoscale}, \code{reso} is ignored.}
}
\description{
Calculate mean of raster objects
Overriding for \code{\link[terra:summarize-generics]{terra::mean()}}. Calculates mean of list of rasters.
}
