% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{ccri_powerlaw}
\alias{ccri_powerlaw}
\title{Calculate inverse power law}
\usage{
ccri_powerlaw(
  dispersal_parameter_beta_vals,
  link_threshold = 0,
  metrics = the$parameters_config$`CCRI parameters`$NetworkMetrics$InversePowerLaw,
  crop_cells_above_threshold = NULL,
  thresholded_crop_values = NULL
)
}
\arguments{
\item{dispersal_parameter_beta_vals}{A list of beta values}

\item{link_threshold}{A threshold value for links.}

\item{metrics}{A list 2 vectors - metrics and weights.}

\item{crop_cells_above_threshold}{A list of crop cells above threshold}

\item{thresholded_crop_values}{A list of crop values above threshold}
}
\value{
A list of calculated inverse power law
}
\description{
Calculate inverse power law
}
\details{
Network metrics should be passed as a list of vectors e.g. \code{list(metrics = c("betweeness"), weights = c(100))}.
Default values are fetched from \code{parameters.yaml} and arguments uses the same structure.
}
