% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{ccri_variance}
\alias{ccri_variance}
\title{Calculate variance of CCRI}
\usage{
ccri_variance(
  indexes,
  rast,
  global,
  geoscale,
  res = reso(),
  outdir = tempdir()
)
}
\arguments{
\item{indexes}{list of rasters. See details.}

\item{rast}{A raster object. It will be used in calculating variance.}

\item{global}{logical. \code{TRUE} if global analysis is required, \code{FALSE} otherwise.
When \code{TRUE}, \code{geoscale} is ignored. Default is \code{TRUE}.}

\item{geoscale}{vector. geographical scale}

\item{res}{numeric. map resolution.}

\item{outdir}{Character. Output directory for saving raster in TIFF format.
Default is \code{\link[=tempdir]{tempdir()}}.}
}
\value{
Invisible NULL.
}
\description{
This function produces a map of variance of CCRI based on input parameters
}
