% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{ccri_variance}
\alias{ccri_variance}
\title{Calculate variance of CCRI
This function produces a map of variance of CCRI based on inpt parameters}
\usage{
ccri_variance(
  indexes,
  variance_mean_index_raster,
  zero_extent_raster,
  map_grey_background_ext,
  resolution = the$parameters_config$`CCRI parameters`$Resolution
)
}
\arguments{
\item{indexes}{A list of index values}

\item{variance_mean_index_raster}{A raster object for variance mean index raster}

\item{zero_extent_raster}{A raster object for zero extent raster}

\item{map_grey_background_ext}{A raster object for map grey background extent}

\item{resolution}{resolution to plot raster and map}
}
\description{
Calculate variance of CCRI
This function produces a map of variance of CCRI based on inpt parameters
}
