% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{ccri_variance}
\alias{ccri_variance}
\title{Calculate variance of CCRI}
\usage{
ccri_variance(
  indexes,
  rast,
  geoscale,
  resolution = the$parameters_config$`CCRI parameters`$Resolution
)
}
\arguments{
\item{indexes}{A list of index values}

\item{rast}{A raster object. It will be used in calculating variance.}

\item{geoscale}{Geographical scale. This will be used in calculating difference.}

\item{resolution}{resolution to plot raster and map}
}
\description{
This function produces a map of variance of CCRI based on input parameters
}
