% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{connectivity}
\alias{connectivity}
\title{Calculate and plot maps}
\usage{
connectivity(
  indexes,
  global = TRUE,
  geoscale,
  reso = reso(),
  pmean = TRUE,
  pvar = TRUE,
  pdiff = TRUE
)
}
\arguments{
\item{indexes}{list of rasters}

\item{geoscale}{vector. geographical scale}

\item{reso}{numeric, map resolution}

\item{pmean}{\code{TRUE} if map of mean should be plotted, \code{FALSE} otherwise.}

\item{pvar}{\code{TRUE} if variance map should be plotted, \code{FALSE} otherwise}

\item{pdiff}{\code{TRUE} if difference map should be plotted, \code{FALSE} otherwise}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Calculate mean, variance and difference. The result is produced in form of maps plotted with predefined settings.
Currently, the settings for plot cannot be customized.
Default value is `TRUE` for all logical arguments
}\if{html}{\out{</div>}}
}
\details{
It will save all the opted plots using - \code{pmean}, \code{pvar} and \code{pdiff}.
File will be saved in \code{\link[=getwd]{getwd()}}.If \code{\link[=interactive]{interactive()}} is \code{TRUE},
then plots can be seen in active plot window. E.g. Rstudio
}
