% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_network.R
\docType{class}
\name{GeoNetwork-class}
\alias{GeoNetwork-class}
\alias{host_density<-,GeoNetwork-method}
\title{GeoNetwork}
\usage{
\S4method{host_density}{GeoNetwork}(x) <- value
}
\arguments{
\item{x}{GeoNetwork.}

\item{value}{SpatRaster.}
}
\value{
GeoNetwork.
}
\description{
An S4 class representing a network of geographical data.
This will wrap all the results from the risk analysis using \code{\link[=sean]{sean()}} or \code{\link[=sensitivity_analysis]{sensitivity_analysis()}}.
This class contains the field from \code{Gmap} class which has results in the form of \code{SpatRaster} and TIFF file.
}
\section{Slots}{

\describe{
\item{\code{rasters}}{A list of \code{GeoRasters} objects.}

\item{\code{host_density}}{A \code{SpatRaster} representing the host density.}

\item{\code{me_rast}}{A raster representing mean risk index.}

\item{\code{me_out}}{Character. A file path to the mean risk index raster.}

\item{\code{diff_rast}}{A raster representing difference.}

\item{\code{diff_out}}{Character. A file path to the difference raster.}

\item{\code{var_rast}}{A raster representing variance.}

\item{\code{var_out}}{Character. A file path to the variance raster.}
}}

