% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_rasters.R
\docType{class}
\name{GeoRasters-class}
\alias{GeoRasters-class}
\alias{.rast_ro}
\title{GeoRaster class}
\description{
A class to represent raster vis-a-vis risk indices.
This class encapsulates the results of apply dispersal models and metrics.
}
\section{Fields}{

\describe{
\item{\code{host_density}}{SpatRaster. A spatial raster representing host density.}

\item{\code{rasters}}{List. List of raster representing risk indices. These are of type \code{GeoModels}.}

\item{\code{global}}{Boolean. True if contains \code{GlobalRast} object, False otherwise.}
}}


