% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{geoscale_param}
\alias{geoscale_param}
\title{Get geographical scales from the parameters}
\usage{
geoscale_param()
}
\value{
A list of geographical scales
}
\description{
This function returns a list of geographical scales set in global and custom extent in parameters.yaml
if \code{global} is \code{TRUE}, the \code{CustomExt} is ignored.
}
