% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rasters.R
\name{get_crop_raster_fromtif}
\alias{get_crop_raster_fromtif}
\title{Get raster object from tif file}
\usage{
get_crop_raster_fromtif(path_to_tif)
}
\arguments{
\item{path_to_tif}{TIF file}
}
\value{
Raster object
}
\description{
This is a wrapper of \code{\link[raster:raster]{raster::raster()}} and generates a raster object if provided with a TIF file.
}
\examples{
\dontrun{
# Generate raster for usage
fp <- .get_helper_filepath("avocado")
get_crop_raster_fromtif(fp)
get_crop_raster_fromtif("avocado_HarvestedAreaFraction.tif")

}
}
