% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{get_crop_raster_fromtif}
\alias{get_crop_raster_fromtif}
\title{Get raster object from tif file
This is wrapper of \code{\link[raster:raster]{raster::raster()}} and generates raster object if provided with tif file.
@param path_to_tif TIF file
@return Raster object
@examples
\dontrun{
 Generate raster for usage
 get_crop_raster_fromtif(system.file("avocado_HarvestedAreaFraction.tif", "tifs",
 package = "geohabnet", mustWork = TRUE))
 }}
\usage{
get_crop_raster_fromtif(path_to_tif)
}
\description{
Get raster object from tif file
This is wrapper of \code{\link[raster:raster]{raster::raster()}} and generates raster object if provided with tif file.
@param path_to_tif TIF file
@return Raster object
@examples
\dontrun{
 Generate raster for usage
 get_crop_raster_fromtif(system.file("avocado_HarvestedAreaFraction.tif", "tifs",
 package = "geohabnet", mustWork = TRUE))
 }
}
