% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rasters.R
\name{get_cropharvest_raster}
\alias{get_cropharvest_raster}
\title{Get raster object for crop}
\usage{
get_cropharvest_raster(crop_name, data_source)
}
\arguments{
\item{crop_name}{Name of the crop}

\item{data_source}{Data source for cropland information}
}
\value{
Raster object
}
\description{
Get cropland information in a form of raster object from data source for crop
}
\examples{
get_cropharvest_raster("avocado", "monfreda")
}
