% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{get_cropharvest_raster}
\alias{get_cropharvest_raster}
\title{Get raster object for crop}
\usage{
get_cropharvest_raster(crop_name)
}
\arguments{
\item{crop_name}{Name of the crop}
}
\value{
Raster object
}
\description{
Takes crop name and returns raster object. Currently,
only supports crops listed in geodata::monfredaCrops()
}
\examples{
get_cropharvest_raster("avocado")
}
