% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{get_cropharvest_raster_sum}
\alias{get_cropharvest_raster_sum}
\title{Takes crop names and returns sum of their rasters}
\usage{
get_cropharvest_raster_sum(crop_names)
}
\arguments{
\item{crop_names}{A vector of crop names}
}
\value{
Raster object
}
\description{
Takes crop names and returns sum of their rasters
}
\examples{
get_cropharvest_raster_sum(c("avocado", "barley"))
}
