% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{get_cropharvest_raster_sum}
\alias{get_cropharvest_raster_sum}
\title{Get sum of rasters for individual crops}
\usage{
get_cropharvest_raster_sum(crop_names)
}
\arguments{
\item{crop_names}{A vector of crop names}
}
\value{
Raster object
}
\description{
Takes crop names and returns raster object which is sum of raster of individual crops.
Currently, only supports crops listed in geodata::monfredaCrops()
}
\examples{
get_cropharvest_raster_sum(c("avocado", "barley"))
}
