% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{get_parameters}
\alias{get_parameters}
\title{Get Parameters}
\usage{
get_parameters(iwindow = FALSE, out_path = getwd())
}
\arguments{
\item{iwindow}{Logical. If TRUE, prompts the user to select the output directory using a file chooser window.}

\item{out_path}{Character. The output path where the parameter file will be copied.}
}
\value{
Character. The path to the copied parameter file.
}
\description{
Retrieves the parameters and copies the parameter file to the specified output path.
}
