% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rasters.R
\name{get_rasters}
\alias{get_rasters}
\title{Get rasters object from parameters
See host object in \code{\link[=get_parameters]{get_parameters()}} or \code{\link[=load_parameters]{load_parameters()}}.}
\usage{
get_rasters(host)
}
\arguments{
\item{host}{SpatRaster. It is synonym to Hosts object in parameters.
This is a wrapper to \code{\link[terra:rast]{terra::rast()}} and generates a raster object if provided with a TIF file.}
}
\value{
List of SpatRaster.
}
\description{
Get rasters object from parameters
See host object in \code{\link[=get_parameters]{get_parameters()}} or \code{\link[=load_parameters]{load_parameters()}}.
}
\examples{
# Get default rasters
\dontrun{
get_rasters("some_raster.tif")
}
}
\seealso{
\code{\link[=load_parameters]{load_parameters()}}, \code{\link[=get_parameters]{get_parameters()}}
}
