% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccri_helper.R
\name{get_supported_sources}
\alias{get_supported_sources}
\title{Get supported sources of crops
When provided, \code{\link[=get_cropharvest_raster]{get_cropharvest_raster()}} will
look for cropland data in this specific source.}
\usage{
get_supported_sources()
}
\value{
return vector of supported sources.
Also used as a lookup to find get raster object.
}
\description{
Get supported sources of crops
When provided, \code{\link[=get_cropharvest_raster]{get_cropharvest_raster()}} will
look for cropland data in this specific source.
}
\examples{
# Get currently supported sources
get_supported_sources()
}
