% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{global_scales<-}
\alias{global_scales<-}
\title{Set global geographical extent}
\usage{
global_scales(east, west) <- value
}
\description{
Set the geographical extents used in global analysis.
Each extent should be in the form of c(Xmin, Xmax, Ymin, Ymax)
@param east Vector. The eastern hemisphere extent.
@param west Vector. The western hemisphere extent.
@export
@seealso \code{\link[terra:ext]{terra::ext()}}
}
