% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{global_scales}
\alias{global_scales}
\title{Global geographical extent}
\usage{
global_scales()
}
\value{
List. Named list with scales for eastern and western hemispheres
}
\description{
This function provides the geographical extent used in a global analysis.
This function returns the geographic extents of the eastern and western hemispheres used in the analysis.
Each geographic extent is in the form of c(Xmin, Xmax, Ymin, Ymax). The geohabnet functions are designed to work with the coordinate reference system: lon/lat WGS 84 (EPSG:4326).
}
\details{
When a habitat connectivity analysis is global, the functions in geohabnet will conduct two separate analyses, one on the geographical scale of the eastern hemisphere and another for the western hemisphere.
The final outcomes (such as maps or adjacency matrices) are then combined for the global analysis.
}
\seealso{
\code{\link[=set_global_scales]{set_global_scales()}}
}
