% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_network.R
\docType{class}
\name{Gmap-class}
\alias{Gmap-class}
\alias{setmaps}
\title{Gmap class}
\usage{
setmaps(x, me, vari, dif)
}
\arguments{
\item{x}{Gmap object.}

\item{me}{SpatRaster. A raster representing mean risk index.}

\item{vari}{SpatRaster. A raster representing variance.}

\item{dif}{SpatRaster. A raster representing difference.}
}
\value{
Gmap object.
}
\description{
An S4 class to represent various maps in the form of \code{SpatRaster}.
}
\section{Slots}{

\describe{
\item{\code{me_rast}}{A raster representing mean risk index.}

\item{\code{me_out}}{Character. A file path to the mean risk index raster.}

\item{\code{diff_rast}}{A raster representing difference.}

\item{\code{diff_out}}{Character. A file path to the difference raster.}

\item{\code{var_rast}}{A raster representing variance.}

\item{\code{var_out}}{Character. A file path to the variance raster.}
}}

