% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_network.R
\docType{class}
\name{Gmap-class}
\alias{Gmap-class}
\alias{setmaps}
\alias{setmaps,Gmap-method}
\title{Gmap class}
\usage{
setmaps(x, me, vari, dif)

\S4method{setmaps}{Gmap}(x, me, vari, dif)
}
\arguments{
\item{x}{A Gmap object.}

\item{me}{A GeoRaster object representing mean risk index.}

\item{vari}{A GeoRaster object representing variance.}

\item{dif}{A GeoRaster object representing difference.}
}
\value{
A Gmap object.
}
\description{
An S4 class to represent various maps.

Set the slots in the Gmap object.
}
\section{Slots}{

\describe{
\item{\code{me_rast}}{SpatRaster A raster representing mean risk index.}

\item{\code{me_out}}{Character. A file path to the mean risk index raster.}

\item{\code{diff_rast}}{SpatRaster A raster representing difference.}

\item{\code{diff_out}}{Character. A file path to the difference raster.}

\item{\code{var_rast}}{Numeric. A raster representing variance.}

\item{\code{var_out}}{SpatRaster A file path to the variance raster.}
}}

