% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{load_parameters}
\alias{load_parameters}
\title{Load Parameters from YAML File}
\usage{
load_parameters(filepath = .get_helper_filepath(.kparameters_file_type))
}
\arguments{
\item{filepath}{Path to the YAML file containing the parameters. By default, it
takes the value of ".kparameters_file_type" which is set to "parameters.yaml".}
}
\value{
object with parameters and values
}
\description{
This function loads parameters from a YAML file and stores them in an object.
}
\examples{
# Load parameters from default file
load_parameters()

}
