% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_density.R
\name{ndplot,GeoNetwork-method}
\alias{ndplot,GeoNetwork-method}
\title{Network density}
\usage{
\S4method{ndplot}{GeoNetwork}(x)
}
\arguments{
\item{x}{A GeoNetwork object}
}
\value{
Vector. Up to two ggplot2 objects
}
\description{
Calculates and plots the network density of a GeoNetwork object.
The plot is produced as dispersal parameter vs edge density.
}
