% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_network.R
\name{ndplot}
\alias{ndplot}
\title{Network density}
\usage{
ndplot(x)
}
\arguments{
\item{x}{A GeoNetwork object}
}
\value{
Vector. Up to two ggplot2 objects
}
\description{
Calculates and plots the network density of a GeoNetwork object.
}
