% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccri_helper.R
\name{risk_indices}
\alias{risk_indices}
\title{Get risk indices}
\usage{
risk_indices(ri)
}
\arguments{
\item{ri}{GeoRasters object}
}
\value{
List of risk indices. If the \code{ri} is global, the list will contain two elements,
one for each hemisphere.
}
\description{
Get risk indices from GeoRasters object.
}
