% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskmap.R
\docType{class}
\name{RiskMap-class}
\alias{RiskMap-class}
\title{RiskMap class}
\description{
An S4 class representing resulting maps from the specific operation type.
}
\section{Fields}{

\describe{
\item{\code{map}}{Character. A file path to the map.}

\item{\code{riid}}{SpatRaster. This is one of the maps of habitat connectivity.}

\item{\code{spr}}{SpatRaster. A spatial raster representing the habitat connectivity index.}

\item{\code{fp}}{Character. A file path to the habitat connectivity raster.}
}}

