% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sean.R
\name{sa_onrasters}
\alias{sa_onrasters}
\title{Run analysis}
\usage{
sa_onrasters(
  rast,
  global = TRUE,
  geo_scale,
  link_thresholds,
  host_density_thresholds,
  aggregate_methods = c("sum", "mean"),
  resolution
)
}
\arguments{
\item{rast}{Raster object which will be used in analysis.}

\item{global}{Logical. \code{TRUE} if global analysis, \code{FALSE} otherwise. Default is \code{TRUE}
Default is \code{TRUE}. when \code{TRUE}, \code{geoscale} is ignored}

\item{geo_scale}{Vector. Geographical scale which is coordinates in the form of c(Xmin, Xmax, Ymin, Ymax)}
}
\description{
Run analysis
}
\examples{
rr <- get_rasters(list(monfreda = c("coffee")))
sa_onrasters(rr[[1]],
            geo_scales = list(c(-115, -75, 5, 32)),
            c(0.0001, 0.00004),
            c(0.0001, 0.00005),
            c("sum", "mean"),
            resolution = 12)
sa_onrasters(rr[[1]],
            geo_scales = list(c(-115, -75, 5, 32)),
            0.0001,
            0.00001,
            c("sum"),
            resolution = 12)

}
\seealso{
Use \code{\link[=get_rasters]{get_rasters()}} to obtain raster object.
}
