% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{sa_onrasters}
\alias{sa_onrasters}
\title{Run analysis}
\usage{
sa_onrasters(
  cropharvest_raster,
  geo_scales,
  link_thresholds,
  host_density_thresholds,
  aggregate_methods = c("sum", "mean"),
  resolution
)
}
\arguments{
\item{cropharvest_raster}{Raster object which will be used in analysis}

\item{geo_scales}{List of geographical scales to be used in analysis.
The rasters will be cropped to provided geographical scale.
Independent analysis is run on each sale.}

\item{link_thresholds}{A list of threshold values for link}

\item{host_density_thresholds}{A list of host density threshold values}

\item{aggregate_methods}{A list of aggregation methods}

\item{resolution}{resolution to plot raster and map}
}
\value{
A list of calculated CCRI values using negative exponential
}
\description{
Run analysis
}
