% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sean.R
\name{sean}
\alias{sean}
\title{Calculate sensitivity analysis on cropland harvested area fraction}
\usage{
sean(
  link_threshold = 0,
  global = TRUE,
  geoscale,
  agg_methods = c("sum", "mean"),
  dist_method = "distGeo",
  rast,
  host_density_threshold = 0,
  reso = reso(),
  maps = TRUE
)
}
\arguments{
\item{link_threshold}{numeric. A threshold value for link}

\item{global}{Logical. \code{TRUE} if global analysis, \code{FALSE} otherwise.
Default is \code{TRUE}}

\item{dist_method}{character. Distance method from \code{\link[=dist_methods]{dist_methods()}}}

\item{rast}{Raster object which will be used in analysis.}

\item{host_density_threshold}{A host density threshold value}

\item{reso}{numeric.
resolution at which operations will run.
Default is \code{\link[=reso]{reso()}}}

\item{maps}{logical. \code{TRUE} if maps are to be plotted, \code{FALSE} otherwise}
}
\value{
A list of calculated CCRI values using negative exponential
}
\description{
This function calculates sensitivity analysis on cropland harvested area fraction based on provided parameters.
It can be used as an entry point for sensitivity analysis.
}
\details{
When \code{global = TRUE}, \code{geoscale} is ignored and \code{\link[=global_scales]{global_scales()}} is used
}
\seealso{
Uses \code{\link[=connectivity]{connectivity()}}
}
