% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sean.R
\name{sean}
\alias{sean}
\alias{msean}
\title{Calculate sensitivity analysis on cropland harvested area fraction}
\usage{
sean(
  rast,
  global = TRUE,
  geoscale = NULL,
  agg_methods = c("sum", "mean"),
  dist_method = "geodesic",
  link_threshold = 0,
  host_density_threshold = 0,
  res = reso()
)

msean(..., global = TRUE, geoscale = NULL, res = reso(), outdir = tempdir())
}
\arguments{
\item{rast}{Raster object which will be used in analysis.}

\item{global}{Logical. \code{TRUE} if global analysis, \code{FALSE} otherwise.
Default is \code{TRUE}}

\item{geoscale}{Numeric vector. Geographical coordinates
in the form of c(Xmin, Xmax, Ymin, Ymax)}

\item{agg_methods}{vector. Aggregation methods}

\item{dist_method}{Character. One of the values from \code{\link[=dist_methods]{dist_methods()}}}

\item{link_threshold}{Numeric. A threshold value for link}

\item{host_density_threshold}{Numeric. A host density threshold value}

\item{res}{Numeric.
resolution at which operations will run.
Default is \code{\link[=reso]{reso()}}}

\item{...}{arguments passed to \code{\link[=sean]{sean()}}}

\item{outdir}{Character. Output directory for saving raster in TIFF format.
Default is \code{\link[=tempdir]{tempdir()}}.}
}
\value{
GeoRasters.

GeoNetwork.
}
\description{
This function calculates sensitivity analysis on cropland harvested area fraction based on provided parameters.
Some parameters are only accessible from \code{parameters.yaml} and uses value from here.
\code{\link[=sensitivity_analysis]{sensitivity_analysis()}} is a wrapper around \code{\link[=sean]{sean()}} function.
\itemize{
\item \code{msean()} is a wrapper around \code{\link[=sean]{sean()}} function. It has additional argument to specify maps which are calculated
using \code{\link[=connectivity]{connectivity()}} function. The maps are essentially the risk network.
}
}
\details{
When \code{global = TRUE}, \code{geoscale} is ignored and \code{\link[=global_scales]{global_scales()}} is used. What makes \code{\link[=sean]{sean()}}
different from \code{\link[=msean]{msean()}} is thier return value. The return value of \code{\link[=msean]{msean()}} is \code{GeoNetwork}
contains the result from applying \code{\link[=connectivity]{connectivity()}} function on the risk indexes. Essentially, the risk maps.
}
\examples{
\donttest{
avocado <- cropharvest_rast("avocado", "monfreda")

# global
ri <- sean(avocado) # returns a list of GeoRasters
mri <- msean(rast = avocado) # returns GeoNetwork object

# non-global
# geoscale is a vector of xmin, xmax, ymin, ymax

# returns GeoRasters object
ri <- sean(avocado, global = FALSE, geoscale = c(-115, -75, 5, 32))
ri

# returns GeoNetwork object
mri <- msean(rast = avocado, global = FALSE, geoscale = c(-115, -75, 5, 32))
mri
}
}
\references{
Yanru Xing, John F Hernandez Nopsa, Kelsey F Andersen, Jorge L Andrade-Piedra, Fenton D Beed,
Guy Blomme, Mónica Carvajal-Yepes, Danny L Coyne, Wilmer J Cuellar, Gregory A Forbes,
Jan F Kreuze, Jürgen Kroschel, P Lava Kumar, James P Legg, Monica Parker, Elmar Schulte-Geldermann,
Kalpana Sharma, Karen A Garrett,
\emph{Global Cropland .connectivity: A Risk Factor for Invasion and Saturation by Emerging Pathogens and Pests},
BioScience, Volume 70, Issue 9, September 2020, Pages 744–758,
\doi{10.1093/biosci/biaa067}

Hijmans R (2023). \emph{terra: Spatial Data Analysis}.
R package version 1.7-46, \url{https://CRAN.R-project.org/package=terra}
}
\seealso{
Uses \code{\link[=connectivity]{connectivity()}}

Uses \code{\link[=msean]{msean()}}
}
