% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sean.R
\name{sean}
\alias{sean}
\title{Calculate sensitivity analysis on cropland harvested area fraction}
\usage{
sean(
  link_threshold = 0,
  global = TRUE,
  geoscale,
  aggregate_methods = c("sum", "mean"),
  rast,
  host_density_threshold = 0,
  resolution = 24
)
}
\arguments{
\item{link_threshold}{A threshold value for link}

\item{global}{Logical. \code{TRUE} if global analysis, \code{FALSE} otherwise. Default is \code{TRUE}}

\item{geoscale}{A geographical extent in form of (Xmin, Xmax, Ymin, Ymax)}

\item{aggregate_methods}{A list of aggregation methods. It can be sum or mean}

\item{host_density_threshold}{A host density threshold value}

\item{resolution}{resolution to plot raster and map}

\item{cropharvest_raster}{A raster object for cropland harvest}
}
\value{
A list of calculated CCRI values using negative exponential
}
\description{
This function calculates sensitivity analysis on cropland harvested area fraction based on provided parameters.
It can be used as an entry point for sensitivity analysis.
}
\details{
When \code{global = TRUE}, geoscale is ignored. Instead global_scales() is used
}
\seealso{
Uses \code{\link[=plot_maps]{plot_maps()}}
}
