% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCRI_v2.R
\name{senstivity_analysis}
\alias{senstivity_analysis}
\title{Calculate sensitivity analysis on parameters}
\usage{
senstivity_analysis()
}
\description{
This function runs sensitivity analysis on parameters based on provided parameters through \code{\link[=set_parameters]{set_parameters()}}.
It can be used as entry point for sensitivity analysis.
Plots results of sensitivity analysis.
}
\examples{
# Run senstivity analysis on default set of parameters
senstivity_analysis()
}
