% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_network.R
\name{setmaps,Gmap-method}
\alias{setmaps,Gmap-method}
\title{Sets the map slots in the Gmap object.}
\usage{
\S4method{setmaps}{Gmap}(x, me, vari, dif)
}
\arguments{
\item{x}{A Gmap object.}

\item{me}{A GeoRaster object representing mean risk index.}

\item{vari}{A GeoRaster object representing variance.}

\item{dif}{A GeoRaster object representing difference.}
}
\value{
A Gmap object.
}
\description{
This wraps the results(SpatRasters) from the risk analysis.
}
