% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{setprops}
\alias{setprops}
\title{Set properties of the GeoModel object.}
\usage{
setprops(x, aggregation, hdthreshold, linkthreshold)
}
\arguments{
\item{x}{The GeoModel object.}

\item{aggregation}{Character. A value representing the aggregation method.}

\item{hdthreshold}{Numeric. A value representing the host density threshold.}

\item{linkthreshold}{Numeric. A value representing the link threshold in a network.}
}
\value{
The GeoModel object with updated properties.
}
\description{
Set properties of the GeoModel object.
}
