% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam_crop.R
\name{sp_rast}
\alias{sp_rast}
\title{raster for mapspam crop.}
\usage{
sp_rast(crp)
}
\arguments{
\item{crp}{character. name of a crop. Case-insensitive.}
}
\value{
spatRaster
}
\description{
get raster for crop in mapspam dataset
}
\details{
See \code{\link[geodata:crop_spam]{geodata::spamCrops()}} for supported crops.
}
\examples{
\dontrun{
sp_rast("rice")
}
}
\references{
International Food Policy Research Institute, 2020.
Spatially-Disaggregated Crop Production Statistics Data in Africa South of the Sahara for 2017.
<doi: 10.7910/DVN/FSSKBW>, Harvard Dataverse, V2
}
\seealso{
\code{\link[geodata:crop_spam]{geodata::spamCrops()}}
\code{\link[=search_crop]{search_crop()}}
}
