% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam_crop.R
\name{sp_rast}
\alias{sp_rast}
\title{Raster for mapspam crop.}
\usage{
sp_rast(crp, africa = FALSE)
}
\arguments{
\item{crp}{character. name of a crop. Case-insensitive.}

\item{africa}{Fetch the 2017 Mpaspam crop data of Africa. Default is FALSE.}
}
\value{
SpatRaster
}
\description{
get raster for crop in mapspam dataset
}
\details{
See \code{\link[geodata:crop_spam]{geodata::spamCrops()}} for supported crops.
}
\examples{
\donttest{
sp_rast("potato")
sp_rast("potato", TRUE)
}
}
\references{
International Food Policy Research Institute, 2020.
Spatially-Disaggregated Crop Production Statistics Data in Africa South of the Sahara for 2017.
<doi: 10.7910/DVN/FSSKBW>, Harvard Dataverse, V2
}
\seealso{
\code{\link[geodata:crop_spam]{geodata::spamCrops()}}
\code{\link[=search_crop]{search_crop()}}
}
